[TICKET-2137] Refactor code. Move JSX to separate components
