+++ src/GameBoard/GameBoard.tsx
@@ -1,53 +1,26 @@
 import { useMemo, useState } from 'react';
-import Button from 'styles/components/Button';
 import {
   GameContainer,
-  GridContainer,
   GameSection,
   PageContainer,
   ScoreSection,
   SectionTitle,
 } from 'styles/components/Grid';
 import GlobalStyle from 'styles/Global';
-import ItemBox from 'styles/components/ItemBox';
-import { itemColors } from 'styles/pages/HomePage/constants';
-import Heading from 'styles/components/Heading';
 import {
   SectionBody,
   ScoreTableColumn,
-  ScoreTable,
-  ScoreTableRow,
-  ScoreTableCell,
 } from 'styles/pages/HomePage/ScoreTable';
-import { LetterItem, getLetterItems } from './constants';
+import { getLetterItems } from './constants';
+import { ItemsState } from 'types';
+import GameButtons from './GameButtons';
+import ScoreSummary from './ScoreSummary';
+import ScoreList from './ScoreList';
 
 export default function GameBoard() {
-  const [itemsState, setItemsState] = useState<{
-    [key: string]: { count: number; score: number; bonusScore: number };
-  }>({});
+  const [itemsState, setItemsState] = useState<ItemsState>({});
 
   const letterItems = getLetterItems();
-
-  const handleLetterClick = (clikedLetter: LetterItem) => () => {
-    setItemsState(prev => {
-      const newCount = (prev[clikedLetter.letter]?.count || 0) + 1;
-      const newScore =
-        (prev[clikedLetter.letter]?.score || 0) + clikedLetter.points;
-      const newBonusScore = clikedLetter.bonus
-        ? Math.floor(newCount / clikedLetter.bonus.count) *
-          clikedLetter.bonus.points
-        : 0;
-
-      return {
-        ...prev,
-        [clikedLetter.letter]: {
-          count: newCount,
-          score: newScore,
-          bonusScore: newBonusScore,
-        },
-      };
-    });
-  };
 
   const overallScore = useMemo(
     () =>
@@ -67,7 +40,9 @@ export default function GameBoard() {
     [itemsState]
   );
 
-  const handleNewGame = () => {
+  const totalScore = overallScore + bonusScore;
+
+  const handleNewGameClick = () => {
     setItemsState({});
   };
 
@@ -79,120 +54,22 @@ export default function GameBoard() {
           <GameSection>
             <SectionTitle>Kahoot! points</SectionTitle>
             <SectionBody>
-              <GridContainer columns="2, 1fr" repeat gap>
-                {letterItems.map(
-                  (letterItem, index) =>
-                    letterItem && (
-                      <Button
-                        type="button"
-                        key={letterItem.letter}
-                        variant={index % itemColors.length}
-                        onClick={handleLetterClick(letterItem)}
-                        aria-label={`Select letter ${letterItem.letter}`}
-                      >
-                        {letterItem.letter}
-                      </Button>
-                    )
-                )}
-              </GridContainer>
+              <GameButtons
+                letterItems={letterItems}
+                setItemsState={setItemsState}
+              />
             </SectionBody>
           </GameSection>
           <ScoreSection>
             <SectionTitle id="scores-section-title">Player items</SectionTitle>
             <ScoreTableColumn>
-              <ScoreTable aria-describedby="scores-section-title">
-                <thead>
-                  <ScoreTableRow>
-                    <ScoreTableCell as="th" scope="col">
-                      Item
-                    </ScoreTableCell>
-                    <ScoreTableCell as="th" scope="col">
-                      Quantity
-                    </ScoreTableCell>
-                    <ScoreTableCell as="th" scope="col">
-                      Score
-                    </ScoreTableCell>
-                  </ScoreTableRow>
-                </thead>
-                <tbody>
-                  {!Object.keys(itemsState).length ? (
-                    <ScoreTableRow>
-                      <ScoreTableCell colSpan={3}>
-                        Choose your item!
-                      </ScoreTableCell>
-                    </ScoreTableRow>
-                  ) : (
-                    Object.keys(itemsState).map(counterKey => (
-                      <ScoreTableRow key={counterKey}>
-                        <ScoreTableCell>
-                          <ItemBox
-                            variant={
-                              letterItems.findIndex(
-                                item => item.letter === counterKey
-                              ) % itemColors.length
-                            }
-                            aria-label={`Item ${counterKey}`}
-                          >
-                            <span>{counterKey}</span>
-                          </ItemBox>
-                        </ScoreTableCell>
-                        <ScoreTableCell
-                          aria-label={`Quantity: ${itemsState[counterKey].count}`}
-                        >
-                          {itemsState[counterKey].count}
-                        </ScoreTableCell>
-                        <ScoreTableCell
-                          aria-label={`Score: ${
-                            itemsState[counterKey].score +
-                            itemsState[counterKey].bonusScore
-                          }`}
-                        >
-                          {itemsState[counterKey].score +
-                            itemsState[counterKey].bonusScore}
-                        </ScoreTableCell>
-                      </ScoreTableRow>
-                    ))
-                  )}
-                </tbody>
-              </ScoreTable>
+              <ScoreList itemsState={itemsState} letterItems={letterItems} />
             </ScoreTableColumn>
-            <GridContainer
-              columns="auto auto"
-              gap
-              justifyContent="start"
-              padding="lg"
-            >
-              <strong>Bonuses</strong>{' '}
-              <span data-testid="bonus-score">{bonusScore}</span>
-            </GridContainer>
-            <div>
-              <hr />
-            </div>
-            <GridContainer
-              columns="auto auto"
-              gap
-              alignItems="center"
-              padding="lg"
-            >
-              <div>
-                <Heading as="p" level={2} noMargin>
-                  Total
-                </Heading>
-                <Heading data-testid="total-score" as="p" level={3} noMargin>
-                  {overallScore + bonusScore}
-                </Heading>
-              </div>
-              <div>
-                <Button
-                  type="button"
-                  onClick={handleNewGame}
-                  fullWidth
-                  aria-label="Start a new game"
-                >
-                  New game
-                </Button>
-              </div>
-            </GridContainer>
+            <ScoreSummary
+              onNewGameClick={handleNewGameClick}
+              bonusScore={bonusScore}
+              totalScore={totalScore}
+            />
           </ScoreSection>
         </GameContainer>
       </PageContainer>
